ml-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.17
pillow==11.0.0
torch==2.6.0
torchvision==0.21.0
easyocr==1.7.1
transformers==4.46.0
groq==0.11.0
numpy==2.1.3
pydantic==2.9.0
python-dotenv==1.0.1